Remove module recorder.history.modern
